+++ package/__init__.py
@@ -1,2 +1,4 @@
 from . import config
 from . import datasets
+from . import datamodules
+from . import models