+++ package/datasets/__init__.py
@@ -1,2 +1,5 @@
 from . import label_dataset
 from . import sentinel2_60m
+
+from .label_dataset import LabelDataset
+from .sentinel2_60m import Sentinel2_60m