+++ package/poc/custom_lightning_module.py
@@ -1,6 +1,6 @@
 import torch
 import torch.nn.functional as F
-import pytorch_lightning as pl
+import lightning.pytorch as pl
 
 
 class CustomLightningModule(pl.LightningModule):